feat: app.feature에는 기능만 남기고 pages로 파일 분리
