--- src/app.feature/Main/Main.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-import ProductList from "./ProductList";
-import BookmarkList from "./BookmarkList";
-
-const Main = () => {
-  return (
-    <StyledWrapper>
-      <ProductList />
-      <BookmarkList />
-    </StyledWrapper>
-  );
-};
-
-export default Main;
-
-const StyledWrapper = styled.div``;
